Moved CarparkView to new class
- Moved CarparkView
- Small changes to Simulator
 - AbstractView is now really abstract
